Restructuring: integrating grep_engine and grep_pipeline
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5892 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/re/re_analysis.cpp
@@ -439,7 +439,7 @@ void ByteTestComplexity::gatherTests(RE * re) {
     }
 }
 
-size_t byteTestComplexity(RE * re) {
+unsigned byteTestComplexity(RE * re) {
     ByteTestComplexity btc_object;
     btc_object.gatherTests(re);
     return btc_object.equalityTests.count() + btc_object.lessThanTests.count();
